backend/requirements-minimal.txt
# Core Django Dependencies
Django==5.0.2
gunicorn>=20.1.0
django-cors-headers==4.3.1
djangorestframework==3.15.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.1.1

# OpenAI and LLM Dependencies
openai==1.58.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.21

# Vector Database (CPU-only)
chromadb==1.0.20
faiss-cpu==1.12.0

# Document Processing (lightweight)
llama-parse==0.4.4
python-multipart==0.0.12
python-docx==1.1.2
pypdf==6.0.0
PyMuPDF==1.26.4

# Basic Data Processing (minimal ML)
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.7.2

# HTTP and API
requests==2.32.5
httpx==0.28.1
aiohttp==3.12.15

# Text Processing and Search
rank-bm25==0.2.2
sentence-transformers==2.7.0

# Utilities
pydantic==2.11.7
pydantic-settings==2.10.1
click==8.2.1
tqdm==4.67.1
rich==14.1.0
typer==0.17.3
